requirements.txt
Django
djangorestframework
psycopg2-binary
drf-spectacular
